Move parseAndDesugar into the library
This will allow parseAndDesugar to be reused in psc-docs.
