+++ src/Language/PureScript/Docs.hs
@@ -10,3 +10,4 @@ import Language.PureScript.Docs.Types as Docs
 import Language.PureScript.Docs.RenderedCode.Types as Docs
 import Language.PureScript.Docs.RenderedCode.Render as Docs
 import Language.PureScript.Docs.Render as Docs
+import Language.PureScript.Docs.ParseAndDesugar as Docs
